[TypographyScheme] Separate protocol target. (#8798)
Splits the `MDCTypographyScheming` protocol target from the implementation
target. This allows clients to conform to the protocol without depending on
the MDC-iOS implementation (and its dependencies).

Fixes #8445
