+++ components/schemes/Typography/src/Scheming/MaterialTypographyScheme+Scheming.h
@@ -1,4 +1,4 @@
-// Copyright 2018-present the Material Components for iOS authors. All Rights Reserved.
+// Copyright 2019-present the Material Components for iOS authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,6 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import <MaterialComponents/MaterialTypographyScheme+BasicFontScheme.h>
-
-#import "MDCTypographyScheme.h"
+#import "MDCTypographyScheming.h"
